recommendation-node/recommendation/servicers/recommendation.py
import grpc
from ..recommendation_pb2_grpc import RecommendationServiceServicer
from ..recommendation_pb2 import RecommendationRequest, RecommendationResponse

class RecommendationService(RecommendationServiceServicer):
    def GetRecommendations(self, request: RecommendationRequest, context: grpc.ServicerContext) -> RecommendationResponse:
        # Implement your recommendation logic here
        return RecommendationResponse(product_ids=["product1", "product2", "product3"])